[update] install.sh

+++ install.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 
 FORCE=0
+GUI=0
 while getopts f OPT
 do
   case $OPT in
-    "f" ) FORCE=1
+    "f" ) FORCE=1 ;;
+    "g" ) GUI=1 ;;
   esac
 done
 
@@ -14,17 +16,24 @@ if [ $(whoami) = "root" -a $FORCE -ne 1 ]; then
   exit 1
 fi
 
+if [ $FORCE -eq 1 ]; then
+  mkdir -p ~/.config/i3/
+  ln -s -f `pwd`/.rls.sh ~/
+  ln -s -f `pwd`/.bar.sh ~/
+  ln -s -f `pwd`/.battery.sh ~/
+  ln -s -f `pwd`/.gtkrc-2.0 ~/
+  ln -s -f `pwd`/.xprofile ~/
+  ln -s -f `pwd`/.config/i3/config ~/.config/i3/config
+fi
+
 mkdir -p ~/.config/nvim/
-mkdir -p ~/.config/i3/
-
 ln -s -f `pwd`/.zshrc ~/
-ln -s -f `pwd`/.rls.sh ~/
-ln -s -f `pwd`/.bar.sh ~/
-ln -s -f `pwd`/.battery.sh ~/
-ln -s -f `pwd`/.gtkrc-2.0 ~/
-ln -s -f `pwd`/.xprofile ~/
 ln -s -f `pwd`/.config/nvim/init.vim ~/.config/nvim/
-ln -s -f `pwd`/.config/i3/config ~/.config/i3/config
-ln -s -f `pwd`/syncpw /usr/local/bin/syncpw
 
-sudo chsh -s /usr/bin/zsh $USER
+#ln -s -f `pwd`/syncpw /usr/local/bin/syncpw
+
+if [ $(whoami) = "root" ]; then
+  chsh -s /bin/zsh $USER
+else
+  sudo chsh -s /bin/zsh $USER
+fi
